# Improve performance of distinct aggregations