--- a/core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/DistinctAggregationStrategyChooser.java
+++ b/core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/DistinctAggregationStrategyChooser.java
@@ -20,6 +20,7 @@ import io.trino.cost.TaskCountEstimator;
 import io.trino.sql.planner.plan.AggregationNode;
 
 import static io.trino.SystemSessionProperties.getTaskConcurrency;
+import static java.lang.Double.NaN;
 import static java.lang.Double.isNaN;
 import static java.util.Objects.requireNonNull;
 
@@ -29,6 +30,7 @@ import static java.util.Objects.requireNonNull;
 public class DistinctAggregationStrategyChooser
 {
     private static final int MARK_DISTINCT_MAX_OUTPUT_ROW_COUNT_MULTIPLIER = 8;
+    private static final int PRE_AGGREGATE_MAX_OUTPUT_ROW_COUNT_MULTIPLIER = MARK_DISTINCT_MAX_OUTPUT_ROW_COUNT_MULTIPLIER * 8;
 
     private final TaskCountEstimator taskCountEstimator;
 
@@ -44,27 +46,43 @@ public class DistinctAggregationStrategyChooser
 
     public boolean shouldAddMarkDistinct(AggregationNode aggregationNode, Session session, StatsProvider statsProvider)
     {
+        return !canParallelizeSingleStepDistinctAggregation(aggregationNode, session, statsProvider, MARK_DISTINCT_MAX_OUTPUT_ROW_COUNT_MULTIPLIER);
+    }
+
+    public boolean shouldUsePreAggregate(AggregationNode aggregationNode, Session session, StatsProvider statsProvider)
+    {
+        if (canParallelizeSingleStepDistinctAggregation(aggregationNode, session, statsProvider, PRE_AGGREGATE_MAX_OUTPUT_ROW_COUNT_MULTIPLIER)) {
+            return false;
+        }
+
+        // mark-distinct is better than pre-aggregate if the number of group-by keys is bigger than 2
+        // because group-by keys are added to every grouping set and this makes partial aggregation behaves badly
+        return aggregationNode.getGroupingKeys().size() <= 2;
+    }
+
+    private boolean canParallelizeSingleStepDistinctAggregation(AggregationNode aggregationNode, Session session, StatsProvider statsProvider, int maxOutputRowCountMultiplier)
+    {
         if (aggregationNode.getGroupingKeys().isEmpty()) {
             // global distinct aggregation is computed using a single thread. MarkDistinct will help parallelize the execution.
-            return true;
+            return false;
         }
 
         double numberOfDistinctValues = getMinDistinctValueCountEstimate(aggregationNode, statsProvider);
         if (isNaN(numberOfDistinctValues)) {
             // if the estimate is unknown, use MarkDistinct to avoid query failure
-            return true;
+            return false;
         }
 
         int maxNumberOfConcurrentThreadsForAggregation = getMaxNumberOfConcurrentThreadsForAggregation(session);
-        if (numberOfDistinctValues <= MARK_DISTINCT_MAX_OUTPUT_ROW_COUNT_MULTIPLIER * maxNumberOfConcurrentThreadsForAggregation) {
+        if (numberOfDistinctValues <= maxOutputRowCountMultiplier * maxNumberOfConcurrentThreadsForAggregation) {
             // small numberOfDistinctValues reduces the distinct aggregation parallelism, also because the partitioning may be skewed.
             // This makes query to underutilize the cluster CPU but also to possibly concentrate memory on few nodes.
             // MarkDistinct should increase the parallelism at a cost of CPU.
-            return true;
+            return false;
         }
 
-        // can parallelize single-step, and single-step distinct is more efficient than MarkDistinct
-        return false;
+        // can parallelize single-step, and single-step distinct is more efficient than alternatives
+        return true;
     }
 
     private int getMaxNumberOfConcurrentThreadsForAggregation(Session session)
@@ -82,6 +100,6 @@ public class DistinctAggregationStrategyChooser
                 .stream()
                 .filter(symbol -> !isNaN(sourceStats.getSymbolStatistics(symbol).getDistinctValuesCount()))
                 .map(symbol -> sourceStats.getSymbolStatistics(symbol).getDistinctValuesCount())
-                .max(Double::compareTo).orElse(Double.NaN);
+                .max(Double::compareTo).orElse(NaN);
     }
 }