… core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/MultipleDistinctAggregationToMarkDistinct.java
… core/trino-main/src/test/java/io/trino/cost/TestOptimizerConfig.java
… core/trino-main/src/test/java/io/trino/sql/planner/TestLogicalPlanner.java
… core/trino-main/src/test/java/io/trino/sql/planner/assertions/PlanMatchPattern.java
… core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/TestDistinctAggregationStrategyChooser.java
… core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/TestMultipleDistinctAggregationToMarkDistinct.java
… core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/TestOptimizeMixedDistinctAggregations.java
… core/trino-main/src/test/java/io/trino/sql/planner/optimizations/TestAddExchangesPlans.java
… core/trino-main/src/test/java/io/trino/sql/query/TestDistinctAggregationsNoMarkDistinct.java
… docs/src/main/sphinx/admin/properties-optimizer.md
… plugin/trino-base-jdbc/src/test/java/io/trino/plugin/jdbc/BaseJdbcConnectorTest.java
… plugin/trino-mysql/src/test/java/io/trino/plugin/mysql/TestMySqlLegacyConnectorTest.java
… plugin/trino-pinot/src/test/java/io/trino/plugin/pinot/TestPinotConnectorSmokeTest.java
… plugin/trino-redshift/src/test/java/io/trino/plugin/redshift/TestRedshiftConnectorTest.java
… testing/trino-testing/src/main/java/io/trino/testing/AbstractTestAggregations.java
… testing/trino-testing/src/main/java/io/trino/testing/AbstractTestEngineOnlyQueries.java
… testing/trino-tests/src/test/java/io/trino/tests/TestOptimizeMixedDistinctAggregations.java