## Description

Improve performance of distinct aggregations by defining additional strategies based on source properties (for example, NDV).

The strategy to use for multiple distinct aggregations.
`SINGLE_STEP` Computes distinct aggregations in single-step without any pre-aggregations.
This strategy will perform poorly if the number of distinct grouping keys is small.
`MARK_DISTINCT` uses `MarkDistinct` for multiple distinct aggregations
or for mix of distinct and non-distinct aggregations.
`PRE_AGGREGATE` Computes distinct aggregations using a combination of aggregation 
and pre-aggregation steps.
`AUTOMATIC` chooses the strategy automatically.

Single-step strategy is preferred. However, for cases with limited concurrency due to
a small number of distinct grouping keys, it will choose an alternative strategy 
based on input data statistics.


| Strategy | Duration | Query |
| --- | --- | --- |
| MARK_DISTINCT | 5353ms | select count(ss_customer_sk), count(distinct ss_ticket_number) from hive.tpcds_sf1000_orc.store_sales;
| PRE_AGGREGATE | 2468ms | select count(ss_customer_sk), count(distinct ss_ticket_number) from hive.tpcds_sf1000_orc.store_sales;
| MARK_DISTINCT | 10109ms | select count(ss_quantity), count(distinct ss_item_sk), count(distinct ss_store_sk) from hive.tpcds_sf1000_orc.store_sales;
| PRE_AGGREGATE | 8253ms | select count(ss_quantity), count(distinct ss_item_sk), count(distinct ss_store_sk) from hive.tpcds_sf1000_orc.store_sales;


<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or is docs only, and no release notes are required.
( ) Release notes are required. Please propose a release note for me.
(x) Release notes are required, with the following suggested text:

```
# General
* Improve performance of distinct aggregations. ({issue}`21907`)
  Replaced configuration property `optimizer.mark-distinct-strategy` with `optimizer.distinct-aggregations-strategy`. 
  It's allowed values are changed as follows: `NONE -> SINGLE_STEP` and `ALWAYS -> MARK_DISTINCT`. 
  Replaced `optimizer.optimize-mixed-distinct-aggregations` configuration property 
  with a new `PRE_AGGREGATE` option in `optimizer.distinct-aggregations-strategy` .
```