@raunaqmorarka 2024-06-04
Please update the release notes section to specify which system session properties and configs were removed/defunct and what are the new alternatives.